refactor: change test package
