+++ src/test/java/lotto/model/LottoResultsTests.java
@@ -1,4 +1,4 @@
-package lottotests;
+package lotto.model;
 
 import lotto.model.LottoResult;
 import lotto.model.LottoResults;
