+++ src/test/java/lotto/model/wrapper/LottoResultMatchIdTests.java
@@ -1,4 +1,4 @@
-package lottotests;
+package lotto.model.wrapper;
 
 import lotto.model.LottoBonusMatchResult;
 import lotto.model.wrapper.LottoResultMatchId;
