+++ src/test/java/stringaddcalculator/StringAddCalculatorTests.java
@@ -1,10 +1,9 @@
-package stringaddcalculatortests;
+package stringaddcalculator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
-import stringaddcalculator.StringAddCalculator;
 import stringaddcalculator.model.Expression;
 import stringaddcalculator.model.Number;
