+++ src/test/java/stringaddcalculator/model/StringNumberTests.java
@@ -1,4 +1,4 @@
-package stringaddcalculatortests;
+package stringaddcalculator.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
